refactor: reorganize lib files
