+++ packages/waku/src/lib/vite-entries/entry.server.tsx
@@ -6,7 +6,7 @@ import { compress } from 'hono/compress';
 import { serveStatic } from '@hono/node-server/serve-static';
 import path from 'node:path';
 import fs from 'node:fs';
-import { DIST_PUBLIC } from '../builder/constants.js';
+import { DIST_PUBLIC } from '../constants.js';
 import { INTERNAL_setAllEnv } from '../../server.js';
 
 function createApp(app: Hono) {
