+++ packages/waku/src/lib/vite-rsc/deploy/aws-lambda/entry.ts
@@ -6,7 +6,7 @@ import { config } from 'virtual:vite-rsc-waku/config';
 import { serveStatic } from '@hono/node-server/serve-static';
 import path from 'node:path';
 import fs from 'node:fs';
-import { DIST_PUBLIC } from '../../../builder/constants.js';
+import { DIST_PUBLIC } from '../../../constants.js';
 import { INTERNAL_setAllEnv } from '../../../../server.js';
 
 function createApp(app: Hono) {
