+++ packages/waku/src/lib/vite-rsc/deploy/deno/entry.ts
@@ -8,7 +8,7 @@ import { createHonoHandler } from '../../engine.js';
 import { honoEnhancer } from 'virtual:vite-rsc-waku/hono-enhancer';
 import { config } from 'virtual:vite-rsc-waku/config';
 import path from 'node:path';
-import { DIST_PUBLIC } from '../../../builder/constants.js';
+import { DIST_PUBLIC } from '../../../constants.js';
 import { INTERNAL_setAllEnv } from '../../../../server.js';
 
 declare let Deno: any;
