+++ packages/waku/src/lib/vite-rsc/deploy/partykit/plugin.ts
@@ -6,7 +6,7 @@ import {
 import type { Config } from '../../../../config.js';
 import path from 'node:path';
 import { existsSync, writeFileSync } from 'node:fs';
-import { DIST_PUBLIC } from '../../../builder/constants.js';
+import { DIST_PUBLIC } from '../../../constants.js';
 import { fileURLToPath } from 'node:url';
 
 const __dirname = fileURLToPath(new URL('.', import.meta.url));
