+++ packages/waku/src/minimal/client.ts
@@ -15,7 +15,7 @@ import type { ReactNode } from 'react';
 import RSDWClient from 'react-server-dom-webpack/client';
 
 import { createCustomError } from '../lib/utils/custom-errors.js';
-import { encodeRscPath, encodeFuncId } from '../lib/renderers/utils.js';
+import { encodeRscPath, encodeFuncId } from '../lib/utils/rsc-path.js';
 
 const { createFromFetch, encodeReply, createTemporaryReferenceSet } =
   RSDWClient;
