+++ packages/waku/src/server.ts
@@ -4,7 +4,7 @@ export {
   getContextData as unstable_getContextData,
 } from './lib/context.js';
 
-export * as unstable_builderConstants from './lib/builder/constants.js';
+export * as unstable_constants from './lib/constants.js';
 
 // The use of `globalThis` in this file is more or less a hack.
 // It should be revisited with a better solution.
